UPDATE: Move parley logic into functions
