+++ crates/oku_core/src/elements/text/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod text;
+mod parley;
 
 pub use self::text::Text;
 pub use self::text::TextState;
